Update_1_OSINT/requirements.txt
# ===============================
# Core Web Framework
# ===============================
Flask==3.0.3
Werkzeug==3.0.1

# ===============================
# OSINT & API Requests
# ===============================
requests==2.32.3
shodan==1.31.0

# ===============================
# Database
# ===============================
Flask-SQLAlchemy==3.1.1

# ===============================
# Data Handling & Analytics
# ===============================
pandas==2.2.2
numpy==1.26.4

# ===============================
# Reporting
# ===============================
openpyxl==3.1.5
fpdf2==2.7.8

# ===============================
# Environment & Config
# ===============================
python-dotenv==1.0.1

# ===============================
# AI / LLM (Optional)
# ===============================
openai>=1.3.0

# ===============================
# Production Server (Optional)
# ===============================
# gunicorn==22.0.0 # Not supported on Windows
